bar chart fix
